pyproject.toml
[project]
name = "finacial-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=1.3.7",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
]
